requirements.txt
Flask
python-dotenv
openai
tiktoken
pytz
gunicorn
requests
